requirements.txt
streamlit
pymongo
python-dotenv
google-generativeai
reportlab
pandas
numpy